requirements.txt
python-dotenv==0.20.0
python-telegram-bot==13.7
requests==2.27.1
google-cloud-dialogflow==2.14.1
google-cloud-language==2.4.3
google-cloud-api-keys


